Add Network Firewall to deployment (#22)
This restructures the system so that the networkfw is filtering all outbound requests, giving us control over what our systems can talk to.

It also restructures the network stuff so that it's not trying to be so damn clever with cidrsubnet. It turned out to be too confusing, so we are now just setting the subnets explicitly. If we need more address space, we can either make the subnets bigger, or we can (probably better) add an eks2 set of subnets with another /22 in it like the eks subnet is set up now.

Also:  make sure that the system can be bootstrapped from zero still.